+++ lib/spass.rb
@@ -1,25 +1,25 @@
 
 class Generator
-  # Create a Generator using the given word dictionary, returning
-  # only words that match the given regular expression
-  def initialize(dict_path='/usr/share/dict/words', options={})
-    @dict_path = dict_path
-    @dict = read_dict(dict_path, options)
+  attr_reader :dict, :chars, :allowed, :file
+  # Create a Generator using the given word dictionary
+  def initialize(file='/usr/share/dict/words', options={})
+    @file = file
+    @allowed = options[:allowed] || /^[a-z]+$/
+    @chars = (options[:chars] || 10).to_i
+    @dict = read_dict
   end
 
   # Read a word dictionary from the given file, and return an array
   # of elements that match the allowed regex. Raise an exception if
   # the given dictionary file cannot be found.
-  def read_dict(path, options={})
-    allowed = options[:allowed] || /^[a-z]+$/
-    chars = (options[:chars] || 10).to_i
-    if !File.file?(path)
-      raise RuntimeError, "Cannot find dict file: #{path}"
+  def read_dict
+    if !File.file?(@file)
+      raise RuntimeError, "Cannot find dict file: #{@file}"
     end
     dict = []
-    IO.readlines(path).each do |line|
+    IO.readlines(@file).each do |line|
       line.strip!
-      if line.length <= chars && line =~ allowed
+      if line.length <= @chars && line =~ @allowed
         dict << line
       end
     end
@@ -31,19 +31,20 @@ class Generator
     @dict[rand(@dict.count)]
   end
 
-  # Return a random number from 1 to 99
+  # Return a random number from 0 to 9
   def random_number
-    1 + rand(99)
+    rand(10)
   end
 
-  # Generate a passphrase of at least the given length in characters
-  def generate(length, options={})
-    phrase = ''
-    while phrase.length < length + 1 # to account for trailing space
-      phrase += "#{random_word} "
-      phrase += "#{random_number} " if options[:digits]
+  # Generate a passphrase with the given number of words
+  def generate(num_words, options={})
+    words = []
+    num_words.times do
+      word = random_word
+      word += random_number.to_s if options[:digits]
+      words << word
     end
-    return phrase.chomp
+    return words.join(' ')
   end
 end